timeloop controller events
