+++ Assets/Prefabs/Interactable.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ed67b44dbec022e47aa9e5267f45554b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
