+++ Assets/ScriptableObjects/Controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8f1d12a4c986581408307b157a663193
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
