+++ Assets/ScriptableObjects/Controller/TimeLimitControllerSO.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 962a79f424033ff408df9840a81294cb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
