+++ Assets/ScriptableObjects/Controller/TimeLimitControllerScriptableObject.cs
@@ -1,0 +1,19 @@
+using System;
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "TimeLimitControllerSO", menuName = "ScriptableObjects/TimeLimitControllerScriptableObject")]
+public class TimeLimitControllerScriptableObject : ScriptableObject
+{
+    public float defaultMaxTime = 10.0f;
+    public float currentMaxTime;
+
+    private void OnEnable()
+    {
+        currentMaxTime = defaultMaxTime;
+    }
+
+    public void ExtendTimeLimit(float extraTime)
+    {
+        currentMaxTime += extraTime;
+    }
+}
