+++ Assets/ScriptableObjects/Schedule.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 35e3c715e8964bf43a30dcad7aafcf2c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
