+++ Assets/ScriptableObjects/Schedule/ScheduleControllerSO.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c677fe107590081418d8b1cbf8a94621, type: 3}
+  m_Name: ScheduleControllerSO
+  m_EditorClassIdentifier: 
+  scheduledEvents:
+  - {fileID: 11400000, guid: 128cd1cd6981cd649a37fd7c5f6260d2, type: 2}
+  - {fileID: 11400000, guid: ad89c2501c65abb49a9959d383928971, type: 2}
