+++ Assets/ScriptableObjects/Schedule/ScheduleControllerSO.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 83de23a067603d640b6f4b857ceb8ecc
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
