+++ Assets/ScriptableObjects/Schedule/ScheduleControllerScriptableObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c677fe107590081418d8b1cbf8a94621
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
