+++ Assets/ScriptableObjects/Schedule/ScheduleEventSO.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ad89c2501c65abb49a9959d383928971
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
