+++ Assets/ScriptableObjects/Schedule/SpawnCubeScheduleEventSO.asset
@@ -1,0 +1,15 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 059023b88a7034648afb482a0d729126, type: 3}
+  m_Name: SpawnCubeScheduleEventSO
+  m_EditorClassIdentifier: 
+  time: 5
