+++ Assets/ScriptableObjects/Schedule/SpawnCubeScheduleEventSO.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 128cd1cd6981cd649a37fd7c5f6260d2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
