--- Assets/ScriptableObjects/TimeLimitControllerSO.asset
@@ -1,16 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b66d0962d5dfd8f418a389db2cece9af, type: 3}
-  m_Name: TimeLimitControllerSO
-  m_EditorClassIdentifier: 
-  defaultMaxTime: 10
-  currentMaxTime: 0
