--- Assets/ScriptableObjects/TimeLimitControllerScriptableObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b66d0962d5dfd8f418a389db2cece9af
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
