+++ Assets/Scripts/Examples.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 340b67cdec04a584b89e28ea563a1562
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
