+++ Assets/Scripts/Examples/InteractableTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0f9319177ed1042489da0897c73755c2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
