+++ Assets/Scripts/Examples/InteractableTimeExtender.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a0d5b7f78f6408845b3921bfd92ed4e4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
