+++ Assets/Scripts/Examples/TestObjectSpawner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ed14765218f168b4ebb481ec288f3493
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
