--- Assets/Scripts/Puzzles.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c4b2602fff39c9346a17401c36499364
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
