--- Assets/Scripts/Puzzles/TestPuzzle.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b5b4085297400a34287ef9a186558168
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
